Generate valid ssh connection string when not passing a port number
The logic in validating connection strings seems incorrect as there
should not be a trailing `:` if there is no port. A connection string
like this actually fails. I added some additional validation for user
too, but it's fairly easy to find an incorrect string that falls through
the gaps. It should prevent many likely mistakes though.